refactor: GoalTaskStats를 domain/model에서 application/dto로 이동
집계 결과 DTO이므로 아키텍처 규칙에 따라 application/dto/에 배치.
GoalTaskStats → GoalTaskStatsDto로 이름 변경.
