+++ src/main/kotlin/kr/io/team/loop/task/presentation/dataloader/GoalTaskStatsDataLoader.kt
@@ -2,8 +2,8 @@ package kr.io.team.loop.task.presentation.dataloader
 
 import com.netflix.graphql.dgs.DgsDataLoader
 import kr.io.team.loop.common.domain.GoalId
+import kr.io.team.loop.task.application.dto.GoalTaskStatsDto
 import kr.io.team.loop.task.application.service.TaskService
-import kr.io.team.loop.task.domain.model.GoalTaskStats
 import org.dataloader.MappedBatchLoader
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.CompletionStage
@@ -11,8 +11,8 @@ import java.util.concurrent.CompletionStage
 @DgsDataLoader(name = "goalTaskStats")
 class GoalTaskStatsDataLoader(
     private val taskService: TaskService,
-) : MappedBatchLoader<Long, GoalTaskStats> {
-    override fun load(keys: Set<Long>): CompletionStage<Map<Long, GoalTaskStats>> =
+) : MappedBatchLoader<Long, GoalTaskStatsDto> {
+    override fun load(keys: Set<Long>): CompletionStage<Map<Long, GoalTaskStatsDto>> =
         CompletableFuture.supplyAsync {
             val goalIds = keys.map { GoalId(it) }.toSet()
             val stats = taskService.getStatsByGoalIds(goalIds)
